Changed how Lua files are loaded. (Description)
The lua files are now sent to the client and included on both sides
depending on what folder you put them in. There are three folders you
add files to; the client, server, and shared folders. You may also
create one subfolder in each for your plugins and the server will send
them to the client.